+++ packages/common/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @front-monitor/common
 
+## 1.0.1
+
+### Patch Changes
+
+- cli
+
 ## 1.0.0
 
 ### Major Changes