+++ packages/types/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @front-monitor/types
 
+## 1.0.1
+
+### Patch Changes
+
+- cli
+- Updated dependencies
+  - @front-monitor/common@1.0.1
+
 ## 1.0.0
 
 ### Major Changes